3_gold_label_curation/requirements.txt
# ACSES Pilot Study Requirements

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# API interaction
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
python-dotenv>=0.19.0  # For environment variable management
tqdm>=4.62.0  # For progress bars (optional)
